#3682: incorporated review comments
